cvrgu_classbot_replit/requirements.txt
python-telegram-bot==21.6
Flask==3.0.3
